Adds godoc comments
