+++ models/user.go
@@ -6,6 +6,7 @@ import (
 	"time"
 )
 
+// User holds information relating to users that use the application
 type User struct {
 	gorm.Model
 	Email       string
