+++ models/website.go
@@ -2,6 +2,7 @@ package models
 
 import "gorm.io/gorm"
 
+// Website holds information about different websites
 type Website struct {
 	gorm.Model
 	Title       string
