+++ ulid/generate.go
@@ -1,4 +1,4 @@
-package util
+package ulid
 
 import (
 	"github.com/oklog/ulid/v2"
@@ -6,7 +6,8 @@ import (
 	"time"
 )
 
-func GenerateULID() string {
+// Generate a new ULID string
+func Generate() string {
 	entropy := ulid.Monotonic(rand.New(rand.NewSource(time.Now().UnixNano())), 0)
 	res := ulid.MustNew(ulid.Timestamp(time.Now()), entropy)
 	return res.String()
